Evitando eliminar categorias que esten en uso

+++ app/Http/Controllers/CategoryController.php
@@ -60,8 +60,19 @@ class CategoryController extends Controller
     public function destroy(int $category)
     {
         $categoryDelete = Category::find($category);
-        $categoryDelete->delete();
+
+        if ($categoryDelete !== null) {
+
+            if ($categoryDelete->skus->count() > 0 || $categoryDelete->rentalPricings->count() > 0) {
+
+                return back()->with('error', 'No se puede eliminar la categoría porque esta siendo usanda en una referencia o en un plan de prestamos.');
+            }
+
+            $categoryDelete->delete();
+            return back()->with('status', 'Categoría eliminada con éxito');
+        }
+
         // dd($categoryDelete);
-        return back()->with('status', 'Categoría eliminada con éxito');
+
     }
 }
